Finish implementing REST API for to dos

+++ api/src/main/kotlin/com/example/todo/ToDoRepository.kt
@@ -2,4 +2,4 @@ package com.example.todo
 
 import org.springframework.data.repository.CrudRepository
 
-interface ToDoRepository : CrudRepository<ToDo, Int>
+interface ToDoRepository : CrudRepository<ToDo, Long>
